@@ -18,7 +18,6 @@ package credentials
 
 import (
 	"bytes"
-	"io/ioutil"
 	"os"
 	"testing"
 	"text/template"
@@ -84,7 +83,7 @@ func TestCreateAzureCredentialFileOnAzureStackCloud(t *testing.T) {
 }
 
 func withAzureCredentials(t *testing.T) {
-	tempFile, err := ioutil.TempFile("", "azure.toml")
+	tempFile, err := os.CreateTemp("", "azure.toml")
 	assert.NoError(t, err)
 	defer func() {
 		err := os.Remove(tempFile.Name())
@@ -114,7 +113,7 @@ func withAzureCredentials(t *testing.T) {
 	assert.Equal(t, testLocation, creds.Location)
 	assert.Equal(t, testVMType, creds.VMType)
 
-	azureCredentialFileContent, err := ioutil.ReadFile(TempAzureCredentialFilePath)
+	azureCredentialFileContent, err := os.ReadFile(TempAzureCredentialFilePath)
 	assert.NoError(t, err)
 
 	const expectedAzureCredentialFileContent = `
@@ -166,7 +165,7 @@ func withEnvironmentVariables(t *testing.T) {
 	assert.Equal(t, testLocation, creds.Location)
 	assert.Equal(t, testVMType, creds.VMType)
 
-	azureCredentialFileContent, err := ioutil.ReadFile(TempAzureCredentialFilePath)
+	azureCredentialFileContent, err := os.ReadFile(TempAzureCredentialFilePath)
 	assert.NoError(t, err)
 
 	const expectedAzureCredentialFileContent = `
