@@ -72,7 +72,7 @@ func newFakeDriverV2(t *testing.T) (*fakeDriverV2, error) {
 	defer ctrl.Finish()
 
 	driver.cloud = azure.GetTestCloud(ctrl)
-	mounter, err := mounter.NewSafeMounter(driver.useCSIProxyGAInterface)
+	mounter, err := mounter.NewSafeMounter(driver.enableWindowsHostProcess, driver.useCSIProxyGAInterface)
 	if err != nil {
 		return nil, err
 	}
