@@ -34,7 +34,15 @@ func TestStart(t *testing.T) {
 	s := NewNonBlockingGRPCServer()
 	// sleep a while to avoid race condition in unit test
 	time.Sleep(time.Millisecond * 500)
-	s.Start("tcp://127.0.0.1:0", nil, nil, nil, true)
+	s.Start("tcp://127.0.0.1:0", nil, nil, nil, true, false)
+	time.Sleep(time.Millisecond * 500)
+}
+
+func TestStartWithOtelTracing(t *testing.T) {
+	s := NewNonBlockingGRPCServer()
+	// sleep a while to avoid race condition in unit test
+	time.Sleep(time.Millisecond * 500)
+	s.Start("tcp://127.0.0.1:0", nil, nil, nil, true, true)
 	time.Sleep(time.Millisecond * 500)
 }
 
@@ -44,7 +52,7 @@ func TestServe(t *testing.T) {
 	s.wg = sync.WaitGroup{}
 	//need to add one here as the actual also requires one.
 	s.wg.Add(1)
-	s.serve("tcp://127.0.0.1:0", nil, nil, nil, true)
+	s.serve("tcp://127.0.0.1:0", nil, nil, nil, true, false)
 }
 
 func TestWait(t *testing.T) {
