@@ -35,7 +35,7 @@ type FakeSafeMounter struct {
 // NewFakeSafeMounter creates a mount.SafeFormatAndMount instance suitable for use in unit tests.
 func NewFakeSafeMounter() (*mount.SafeFormatAndMount, error) {
 	if runtime.GOOS == "windows" {
-		return NewSafeMounter(true)
+		return NewSafeMounter(true, true)
 	}
 
 	fakeSafeMounter := FakeSafeMounter{}
