@@ -17,7 +17,6 @@ limitations under the License.
 package e2e
 
 import (
-	"context"
 	"fmt"
 
 	"sigs.k8s.io/azuredisk-csi-driver/test/e2e/driver"
@@ -48,7 +47,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 		skipVolumeDeletion bool
 	)
 
-	ginkgo.BeforeEach(func() {
+	ginkgo.BeforeEach(func(ctx ginkgo.SpecContext) {
 		cs = f.ClientSet
 		ns = f.Namespace
 		testDriver = driver.InitAzureDiskDriver()
@@ -56,12 +55,12 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 		skipVolumeDeletion = false
 	})
 
-	ginkgo.AfterEach(func() {
+	ginkgo.AfterEach(func(ctx ginkgo.SpecContext) {
 		if !skipVolumeDeletion {
 			req := &csi.DeleteVolumeRequest{
 				VolumeId: volumeID,
 			}
-			_, err := azurediskDriver.DeleteVolume(context.Background(), req)
+			_, err := azurediskDriver.DeleteVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume %q error: %v", volumeID, err))
 			}
@@ -69,13 +68,13 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 	})
 
 	ginkgo.Context("[single-az]", func() {
-		ginkgo.It("should use a pre-provisioned volume and mount it as readOnly in a pod [disk.csi.azure.com][windows]", func() {
+		ginkgo.It("should use a pre-provisioned volume and mount it as readOnly in a pod [disk.csi.azure.com][windows]", func(ctx ginkgo.SpecContext) {
 			// Az tests need to be changed to pass the right parameters for in-tree driver.
 			// Skip these tests until above is fixed.
 			skipIfUsingInTreeVolumePlugin()
 
 			req := makeCreateVolumeReq("pre-provisioned-readOnly", defaultDiskSize)
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -107,16 +106,16 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 				CSIDriver: testDriver,
 				Pods:      pods,
 			}
-			test.Run(cs, ns)
+			test.Run(ctx, cs, ns)
 		})
 
-		ginkgo.It(fmt.Sprintf("should use a pre-provisioned volume and retain PV with reclaimPolicy %q [disk.csi.azure.com][windows]", v1.PersistentVolumeReclaimRetain), func() {
+		ginkgo.It(fmt.Sprintf("should use a pre-provisioned volume and retain PV with reclaimPolicy %q [disk.csi.azure.com][windows]", v1.PersistentVolumeReclaimRetain), func(ctx ginkgo.SpecContext) {
 			// Az tests need to be changed to pass the right parameters for in-tree driver.
 			// Skip these tests until above is fixed.
 			skipIfUsingInTreeVolumePlugin()
 
 			req := makeCreateVolumeReq("pre-provisioned-retain-reclaimPolicy", defaultDiskSize)
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -139,10 +138,10 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 				Volumes:       volumes,
 				VolumeContext: resp.Volume.VolumeContext,
 			}
-			test.Run(cs, ns)
+			test.Run(ctx, cs, ns)
 		})
 
-		ginkgo.It("should succeed when creating a shared disk [disk.csi.azure.com][windows]", func() {
+		ginkgo.It("should succeed when creating a shared disk [disk.csi.azure.com][windows]", func(ctx ginkgo.SpecContext) {
 			skipIfUsingInTreeVolumePlugin()
 			skipIfOnAzureStackCloud()
 			req := makeCreateVolumeReq("single-shared-disk", 512)
@@ -155,7 +154,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			req.VolumeCapabilities[0].AccessType = &csi.VolumeCapability_Block{
 				Block: &csi.VolumeCapability_BlockVolume{},
 			}
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -163,7 +162,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			ginkgo.By(fmt.Sprintf("Successfully provisioned a shared disk volume: %q\n", volumeID))
 		})
 
-		ginkgo.It("should fail when maxShares is invalid [disk.csi.azure.com][windows]", func() {
+		ginkgo.It("should fail when maxShares is invalid [disk.csi.azure.com][windows]", func(ctx ginkgo.SpecContext) {
 			// Az tests need to be changed to pass the right parameters for in-tree driver.
 			// Skip these tests until above is fixed.
 			skipIfUsingInTreeVolumePlugin()
@@ -171,11 +170,11 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			skipVolumeDeletion = true
 			req := makeCreateVolumeReq("invalid-maxShares", 256)
 			req.Parameters = map[string]string{"maxShares": "0"}
-			_, err := azurediskDriver.CreateVolume(context.Background(), req)
+			_, err := azurediskDriver.CreateVolume(ctx, req)
 			framework.ExpectError(err)
 		})
 
-		ginkgo.It("should succeed when attaching a shared block volume to multiple pods [disk.csi.azure.com][shared disk]", func() {
+		ginkgo.It("should succeed when attaching a shared block volume to multiple pods [disk.csi.azure.com][shared disk]", func(ctx ginkgo.SpecContext) {
 			skipIfUsingInTreeVolumePlugin()
 			skipIfOnAzureStackCloud()
 			skipIfTestingInWindowsCluster()
@@ -195,7 +194,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			req.VolumeCapabilities[0].AccessMode = &csi.VolumeCapability_AccessMode{
 				Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_MULTI_WRITER,
 			}
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -246,10 +245,10 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 				PodCheck:      podCheck,
 				VolumeContext: resp.Volume.VolumeContext,
 			}
-			test.Run(cs, ns)
+			test.Run(ctx, cs, ns)
 		})
 
-		ginkgo.It("should succeed when creating a PremiumV2_LRS disk [disk.csi.azure.com][windows]", func() {
+		ginkgo.It("should succeed when creating a PremiumV2_LRS disk [disk.csi.azure.com][windows]", func(ctx ginkgo.SpecContext) {
 			skipIfUsingInTreeVolumePlugin()
 			skipIfOnAzureStackCloud()
 			req := makeCreateVolumeReq("premium-v2-disk", 100)
@@ -262,7 +261,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			req.VolumeCapabilities[0].AccessType = &csi.VolumeCapability_Block{
 				Block: &csi.VolumeCapability_BlockVolume{},
 			}
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -270,7 +269,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			ginkgo.By(fmt.Sprintf("Successfully provisioned a PremiumV2_LRS disk volume: %q\n", volumeID))
 		})
 
-		ginkgo.It("should succeed when reattaching a disk to a new node on DanglingAttachError [disk.csi.azure.com]", func() {
+		ginkgo.It("should succeed when reattaching a disk to a new node on DanglingAttachError [disk.csi.azure.com]", func(ctx ginkgo.SpecContext) {
 			skipIfUsingInTreeVolumePlugin()
 			skipIfOnAzureStackCloud()
 			req := makeCreateVolumeReq("reattach-disk-multiple-nodes", defaultDiskSize)
@@ -281,7 +280,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 			req.VolumeCapabilities[0].AccessType = &csi.VolumeCapability_Block{
 				Block: &csi.VolumeCapability_BlockVolume{},
 			}
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -313,10 +312,10 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 				Pod:             pod,
 				VolumeContext:   resp.Volume.VolumeContext,
 			}
-			test.Run(cs, ns)
+			test.Run(ctx, cs, ns)
 		})
 
-		ginkgo.It("should create an inline volume by in-tree driver [kubernetes.io/azure-disk]", func() {
+		ginkgo.It("should create an inline volume by in-tree driver [kubernetes.io/azure-disk]", func(ctx ginkgo.SpecContext) {
 			if !isUsingInTreeVolumePlugin {
 				ginkgo.Skip("test case is only available for csi driver")
 			}
@@ -326,7 +325,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 
 			skipVolumeDeletion = true
 			req := makeCreateVolumeReq("pre-provisioned-inline-volume", defaultDiskSize)
-			resp, err := azurediskDriver.CreateVolume(context.Background(), req)
+			resp, err := azurediskDriver.CreateVolume(ctx, req)
 			if err != nil {
 				ginkgo.Fail(fmt.Sprintf("create volume error: %v", err))
 			}
@@ -359,7 +358,7 @@ var _ = ginkgo.Describe("Pre-Provisioned", func() {
 				DiskURI:   volumeID,
 				ReadOnly:  false,
 			}
-			test.Run(cs, ns)
+			test.Run(ctx, cs, ns)
 		})
 	})
 })
